requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
urllib3==2.0.7
